Quick fix for HTTP-based S3 endpoints
The existing code simply provides the IP address of the leader unit of
the target application.

This is unfortunately insufficient; the code in boto runs some
verification of the endpoint address, expecting
urlparse(address).hostname to not return None.  However, if provided
with a bare IP address, urlparse considers this as part of the path
fragment rather than the netloc, and the hostname parameter returns as
None.

If the HTTP scheme is added (e.g. http://), this allows for parsing as
expected.

Note: this patch is a quick fix for the HTTP case.  It does not cover
HTTPS; something more sophisticated will be needed for that.